Reorg existing layout, remove old data-frame file
