+++ ui/core/src/index.ts
@@ -18,11 +18,11 @@ export * from './hooks/datasource';
 export * from './hooks/memo';
 export * from './hooks/panel';
 export * from './model/graph-query';
-export * from './model/data-frame';
 export * from './model/definitions';
 export * from './model/dashboard';
 export * from './model/datasource';
 export * from './model/json-references';
+export * from './model/layout';
 export * from './model/panels';
 export * from './model/plugins';
 export * from './model/resource';
